Add matchesSafely implementation to CompositePropertyMatcher.
The new default implementation invokes all registered property matchers
that have not already been applied. This means that a subclass can either
use the default implementation to auto-match all property matchers, or
override the default implementation and manually apply one or more
matchers, then invoke the default implementation to apply the remaining
property matchers.

Note that the default implementation simply passes the
item to the property matcher, and not the value of the specific property.
In practice, this means that property matchers invoked automatically should
be ReflectingPropertyMatchers, since these determine the correct property
value from the parent object.